% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosspca.R
\name{mapcols}
\alias{mapcols}
\title{Find a mapping of columns from cols into refcols that most preserves similarity.}
\usage{
mapcols(cols, refcols)
}
\arguments{
\item{cols}{matrix}

\item{refcols}{matrix nrow(cols)==nrow(refcol), ncol(cols)>=ncol(refcols)}
}
\value{
mapping perm of cols to refcols s.t. cols[,perm[j]] ~ refcols[,j] j in seq_len(refcols)
}
\description{
Find a mapping of columns from cols into refcols that most preserves similarity.
}
\examples{

x <- USArrests
perm <- c(2,3,4,1)
x2 <- x[,perm]
mp <- crosspca:::mapcols(x2,x) # should be inverse permutation.

}

