% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosspca.R
\name{xprcomp}
\alias{xprcomp}
\title{compute principal projection simulating out of sample data.}
\usage{
xprcomp(x, k, crossplan, center, scale.)
}
\arguments{
\item{x}{matrix to perfrom principal components analysis on.}

\item{k}{integer number of components to estimate.}

\item{crossplan}{list of lists which have train and app slots, where app is a partition of seq_len(nrow(x)).}

\item{center}{logical controls centering of data.}

\item{scale.}{logical controls scaling of data.}
}
\value{
training (simulated out of sample projected) matrix and projection.
}
\description{
compute principal projection simulating out of sample data.
}
\examples{


set.seed(23552)
splitPlan <- vtreat::kWayCrossValidation(nrow(USArrests),3,NULL,NULL)
dc <- xprcomp(USArrests,k=2,crossplan=splitPlan,center=TRUE,scale.=TRUE)

}

